Organize all ci_adapter unit tests in the same folder
We had a `ci_adapters` folder under `tests` with some tests, but tests to other adapters where under `test_ci_adapters.py` directly.
This makes it harder to keep track of which CIs we have tests for and where they are.

This is a simple attempt at putting everything in the same place separatedly.
THe old `test_ci_adapters.py` contains now only the selection tests.
